Some README fixes. Refactor AWSDynamoDBTable to reduce file size. Add travis confiuration.
