+++ Sources/SmokeDynamoDB/AWSDynamoDBTable.swift
@@ -22,14 +22,14 @@ import DynamoDBModel
 import SmokeAWSCore
 import SmokeHTTPClient
 
-private let dynamodbEncoder = DynamoDBEncoder()
-private let dynamodbDecoder = DynamoDBDecoder()
-private let jsonEncoder = JSONEncoder()
-private let jsonDecoder = JSONDecoder()
-
 public class AWSDynamoDBTable: DynamoDBTable {
-    private let dynamodb: AWSDynamoDBClient
-    private let targetTableName: String
+    internal let dynamodb: AWSDynamoDBClient
+    internal let targetTableName: String
+    
+    static internal let dynamodbEncoder = DynamoDBEncoder()
+    static internal let dynamodbDecoder = DynamoDBDecoder()
+    static internal let jsonEncoder = JSONEncoder()
+    static internal let jsonDecoder = JSONDecoder()
     
     private let defaultPaginationLimit = 100
     
@@ -62,19 +62,6 @@ public class AWSDynamoDBTable: DynamoDBTable {
         self.targetTableName = tableName
         
         Log.info("AWSDynamoDBTable created with region '\(region)' and hostname: '\(endpointHostName)'")
-    }
-    
-    private func getInputForInsert<AttributesType, ItemType>(_ item: TypedDatabaseItem<AttributesType, ItemType>) throws
-        -> DynamoDBModel.PutItemInput {
-            let attributes = try getAttributes(forItem: item)
-
-            let expressionAttributeNames = ["#pk": AttributesType.paritionKeyAttributeName, "#sk": AttributesType.sortKeyAttributeName]
-            let conditionExpression = "attribute_not_exists (#pk) AND attribute_not_exists (#sk)"
-            
-            return DynamoDBModel.PutItemInput(conditionExpression: conditionExpression,
-                                              expressionAttributeNames: expressionAttributeNames,
-                                              item: attributes,
-                                              tableName: targetTableName)
     }
     
     public func insertItemSync<AttributesType, ItemType>(_ item: TypedDatabaseItem<AttributesType, ItemType>) throws {
@@ -112,23 +99,6 @@ public class AWSDynamoDBTable: DynamoDBTable {
             try putItemAsync(forInput: putItemInput, withKey: item.compositePrimaryKey, completion: completion)
     }
     
-    private func getInputForUpdateItem<AttributesType, ItemType>(
-            newItem: TypedDatabaseItem<AttributesType, ItemType>,
-            existingItem: TypedDatabaseItem<AttributesType, ItemType>) throws -> DynamoDBModel.PutItemInput {
-        let attributes = try getAttributes(forItem: newItem)
-        
-        let expressionAttributeNames = ["#rowversion": RowStatus.CodingKeys.rowVersion.stringValue]
-        let expressionAttributeValues = [":versionnumber": DynamoDBModel.AttributeValue(N: String(existingItem.rowStatus.rowVersion))]
-        
-        let conditionExpression = "#rowversion = :versionnumber"
-        
-        return DynamoDBModel.PutItemInput(conditionExpression: conditionExpression,
-                                                      expressionAttributeNames: expressionAttributeNames,
-                                                      expressionAttributeValues: expressionAttributeValues,
-                                                      item: attributes,
-                                                      tableName: targetTableName)
-    }
-    
     public func updateItemSync<AttributesType, ItemType>(newItem: TypedDatabaseItem<AttributesType, ItemType>,
                                                          existingItem: TypedDatabaseItem<AttributesType, ItemType>) throws {
         let putItemInput = try getInputForUpdateItem(newItem: newItem, existingItem: existingItem)
@@ -145,18 +115,6 @@ public class AWSDynamoDBTable: DynamoDBTable {
             try putItemAsync(forInput: putItemInput, withKey: newItem.compositePrimaryKey, completion: completion)
     }
     
-    private func getInputForGetItem<AttributesType>(forKey key: CompositePrimaryKey<AttributesType>) throws -> DynamoDBModel.GetItemInput {
-        let attributeValue = try dynamodbEncoder.encode(key)
-        
-        if let keyAttributes = attributeValue.M {
-            return DynamoDBModel.GetItemInput(consistentRead: true,
-                                              key: keyAttributes,
-                                              tableName: targetTableName)
-        } else {
-            throw SmokeDynamoDBError.databaseError(reason: "Expected a structure.")
-        }
-    }
-    
     public func getItemSync<AttributesType, ItemType>(forKey key: CompositePrimaryKey<AttributesType>) throws
         -> TypedDatabaseItem<AttributesType, ItemType>? {
             let putItemInput = try getInputForGetItem(forKey: key)
@@ -167,7 +125,7 @@ public class AWSDynamoDBTable: DynamoDBTable {
             if let item = attributeValue.item {
                 Log.verbose("Value returned from DynamoDB.")
                 
-                return try dynamodbDecoder.decode(DynamoDBModel.AttributeValue(M: item))
+                return try AWSDynamoDBTable.dynamodbDecoder.decode(DynamoDBModel.AttributeValue(M: item))
             } else {
                 Log.verbose("No item returned from DynamoDB.")
                 
@@ -189,7 +147,7 @@ public class AWSDynamoDBTable: DynamoDBTable {
                         
                         do {
                             let decodedItem: TypedDatabaseItem<AttributesType, ItemType>? =
-                                try dynamodbDecoder.decode(DynamoDBModel.AttributeValue(M: item))
+                                try AWSDynamoDBTable.dynamodbDecoder.decode(DynamoDBModel.AttributeValue(M: item))
                             completion(.response(decodedItem))
                         } catch {
                             completion(.error(error))
@@ -203,17 +161,6 @@ public class AWSDynamoDBTable: DynamoDBTable {
                     completion(.error(error))
                 }
             }
-    }
-    
-    private func getInputForDeleteItem<AttributesType>(forKey key: CompositePrimaryKey<AttributesType>) throws -> DynamoDBModel.DeleteItemInput {
-        let attributeValue = try dynamodbEncoder.encode(key)
-        
-        if let keyAttributes = attributeValue.M {
-            return DynamoDBModel.DeleteItemInput(key: keyAttributes,
-                                                 tableName: targetTableName)
-        } else {
-            throw SmokeDynamoDBError.databaseError(reason: "Expected a structure.")
-        }
     }
     
     public func deleteItemSync<AttributesType>(forKey key: CompositePrimaryKey<AttributesType>) throws {
@@ -318,73 +265,6 @@ public class AWSDynamoDBTable: DynamoDBTable {
                           completion: handleQueryResult)
     }
     
-    private func getQueryInput<AttributesType>(forPartitionKey partitionKey: String,
-                                               primaryKeyType: AttributesType.Type,
-                                               sortKeyCondition: AttributeCondition?,
-                                               limit: Int,
-                                               exclusiveStartKey: String?) throws
-        -> DynamoDBModel.QueryInput where AttributesType: PrimaryKeyAttributes {
-        let expressionAttributeValues: [String: DynamoDBModel.AttributeValue]
-        let expressionAttributeNames: [String: String]
-        let keyConditionExpression: String
-        if let currentSortKeyCondition = sortKeyCondition {
-            var withSortConditionAttributeValues: [String: DynamoDBModel.AttributeValue] = [
-                ":pk": DynamoDBModel.AttributeValue(S: partitionKey)]
-            
-            let sortKeyExpression: String
-            switch currentSortKeyCondition {
-            case .equals(let value):
-                withSortConditionAttributeValues[":sortkeyval"] = DynamoDBModel.AttributeValue(S: value)
-                sortKeyExpression = "#sk = :sortkeyval"
-            case .lessThan(let value):
-                withSortConditionAttributeValues[":sortkeyval"] = DynamoDBModel.AttributeValue(S: value)
-                sortKeyExpression = "#sk < :sortkeyval"
-            case .lessThanOrEqual(let value):
-                withSortConditionAttributeValues[":sortkeyval"] = DynamoDBModel.AttributeValue(S: value)
-                sortKeyExpression = "#sk <= :sortkeyval"
-            case .greaterThan(let value):
-                withSortConditionAttributeValues[":sortkeyval"] = DynamoDBModel.AttributeValue(S: value)
-                sortKeyExpression = "#sk > :sortkeyval"
-            case .greaterThanOrEqual(let value):
-                withSortConditionAttributeValues[":sortkeyval"] = DynamoDBModel.AttributeValue(S: value)
-                sortKeyExpression = "#sk >= :sortkeyval"
-            case .between(let value1, let value2):
-                withSortConditionAttributeValues[":sortkeyval1"] = DynamoDBModel.AttributeValue(S: value1)
-                withSortConditionAttributeValues[":sortkeyval2"] = DynamoDBModel.AttributeValue(S: value2)
-                sortKeyExpression = "#sk BETWEEN :sortkeyval1 AND :sortkeyval2"
-            case .beginsWith(let value):
-                withSortConditionAttributeValues[":sortkeyval"] = DynamoDBModel.AttributeValue(S: value)
-                sortKeyExpression = "begins_with ( #sk, :sortkeyval )"
-            }
-            
-            keyConditionExpression = "#pk= :pk AND \(sortKeyExpression)"
-            
-            expressionAttributeNames = ["#pk": AttributesType.paritionKeyAttributeName,
-                                        "#sk": AttributesType.sortKeyAttributeName]
-            expressionAttributeValues = withSortConditionAttributeValues
-        } else {
-            keyConditionExpression = "#pk= :pk"
-            
-            expressionAttributeNames = ["#pk": AttributesType.paritionKeyAttributeName]
-            expressionAttributeValues = [":pk": DynamoDBModel.AttributeValue(S: partitionKey)]
-        }
-            
-        let inputExclusiveStartKey: [String: DynamoDBModel.AttributeValue]?
-        if let exclusiveStartKey = exclusiveStartKey?.data(using: .utf8) {
-            inputExclusiveStartKey = try jsonDecoder.decode([String: DynamoDBModel.AttributeValue].self,
-                                                            from: exclusiveStartKey)
-        } else {
-            inputExclusiveStartKey = nil
-        }
-    
-        return DynamoDBModel.QueryInput(exclusiveStartKey: inputExclusiveStartKey,
-                                        expressionAttributeNames: expressionAttributeNames,
-                                        expressionAttributeValues: expressionAttributeValues,
-                                        keyConditionExpression: keyConditionExpression,
-                                        limit: limit,
-                                        tableName: targetTableName)
-    }
-    
     public func querySync<AttributesType, PossibleTypes>(forPartitionKey partitionKey: String,
                                                          sortKeyCondition: AttributeCondition?,
                                                          limit: Int,
@@ -398,7 +278,7 @@ public class AWSDynamoDBTable: DynamoDBTable {
             
             let lastEvaluatedKey: String?
             if let returnedLastEvaluatedKey = queryOutput.lastEvaluatedKey {
-                let encodedLastEvaluatedKey = try jsonEncoder.encode(returnedLastEvaluatedKey)
+                let encodedLastEvaluatedKey = try AWSDynamoDBTable.jsonEncoder.encode(returnedLastEvaluatedKey)
                 
                 lastEvaluatedKey = String(data: encodedLastEvaluatedKey, encoding: .utf8)
             } else {
@@ -409,7 +289,7 @@ public class AWSDynamoDBTable: DynamoDBTable {
                 let items: [PolymorphicDatabaseItem<AttributesType, PossibleTypes>] = try outputAttributeValues.map { values in
                     let attributeValue = DynamoDBModel.AttributeValue(M: values)
                     
-                    return try dynamodbDecoder.decode(attributeValue)
+                    return try AWSDynamoDBTable.dynamodbDecoder.decode(attributeValue)
                 }
                 
                 return (items, lastEvaluatedKey)
@@ -435,7 +315,7 @@ public class AWSDynamoDBTable: DynamoDBTable {
                         let encodedLastEvaluatedKey: Data
                         
                         do {
-                            encodedLastEvaluatedKey = try jsonEncoder.encode(returnedLastEvaluatedKey)
+                            encodedLastEvaluatedKey = try AWSDynamoDBTable.jsonEncoder.encode(returnedLastEvaluatedKey)
                         } catch {
                             return completion(.error(error))
                         }
@@ -452,7 +332,7 @@ public class AWSDynamoDBTable: DynamoDBTable {
                             items = try outputAttributeValues.map { values in
                                 let attributeValue = DynamoDBModel.AttributeValue(M: values)
                                 
-                                return try dynamodbDecoder.decode(attributeValue)
+                                return try AWSDynamoDBTable.dynamodbDecoder.decode(attributeValue)
                             }
                         } catch {
                             return completion(.error(error))
@@ -466,20 +346,6 @@ public class AWSDynamoDBTable: DynamoDBTable {
                     return completion(.error(error))
                 }
             }
-    }
-    
-    private func getAttributes<AttributesType, ItemType>(forItem item: TypedDatabaseItem<AttributesType, ItemType>) throws
-        -> [String: DynamoDBModel.AttributeValue] {
-            let attributeValue = try dynamodbEncoder.encode(item)
-            
-            let attributes: [String: DynamoDBModel.AttributeValue]
-            if let itemAttributes = attributeValue.M {
-                attributes = itemAttributes
-            } else {
-                throw SmokeDynamoDBError.databaseError(reason: "Expected a map.")
-            }
-            
-            return attributes
     }
     
     private func putItemSync<AttributesType>(forInput putItemInput: DynamoDBModel.PutItemInput,
